VisionVazheh_Backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field[phonenumbers]
django-cors-headers
gtts
Pillow